Chore: 移除地质云授权外的图
